把添加/删除收藏功能移到视频详情页
