+++ app/src/main/java/com/optimalorange/cooltechnologies/ui/ShowVideoDetailActivity.java
@@ -5,9 +5,12 @@ import com.android.volley.VolleyError;
 import com.android.volley.toolbox.ImageLoader;
 import com.optimalorange.cooltechnologies.R;
 import com.optimalorange.cooltechnologies.entity.FavoriteBean;
+import com.optimalorange.cooltechnologies.network.CreateFavoriteRequest;
+import com.optimalorange.cooltechnologies.network.DestroyFavoriteRequest;
 import com.optimalorange.cooltechnologies.network.NetworkChecker;
 import com.optimalorange.cooltechnologies.network.VideoDetailRequest;
 import com.optimalorange.cooltechnologies.network.VolleySingleton;
+import com.optimalorange.cooltechnologies.storage.DefaultSharedPreferencesSingleton;
 import com.optimalorange.cooltechnologies.ui.entity.Video;
 
 import org.json.JSONException;
@@ -17,15 +20,17 @@ import android.app.Activity;
 import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
+import android.view.Menu;
+import android.view.MenuItem;
 import android.view.View;
 import android.widget.ImageView;
 import android.widget.TextView;
+import android.widget.Toast;
 
 import java.lang.ref.WeakReference;
 
-public class ShowVideoDetailActivity extends AppCompatActivity {
+public class ShowVideoDetailActivity extends LoginableBaseActivity {
 
     /**
      * 应当播放的Video的ID<br/>
@@ -35,9 +40,16 @@ public class ShowVideoDetailActivity extends AppCompatActivity {
             ShowVideoDetailActivity.class.getName() + ".extra.KEY_VIDEO_ID";
 
 
+    private VolleySingleton mVolleySingleton;
+
     private String mYoukuClientId;
 
     private Video mVideo;
+
+    /**
+     * 状态属性：已经收藏此视频
+     */
+    private boolean mHasBookmarked = false;
 
     private ViewHolder mViews;
 
@@ -69,8 +81,7 @@ public class ShowVideoDetailActivity extends AppCompatActivity {
             final ImageLoader.ImageListener imageListener = ImageLoader.getImageListener(
                     mViews.thumbnail, 0, 0
             );
-            VolleySingleton.getInstance(mViews.thumbnail.getContext()).getImageLoader()
-                    .get(video.bigThumbnail, imageListener);
+            mVolleySingleton.getImageLoader().get(video.bigThumbnail, imageListener);
         }
     }
 
@@ -94,6 +105,13 @@ public class ShowVideoDetailActivity extends AppCompatActivity {
         return result;
     }
 
+    /**
+     * 有新添加的菜单项
+     */
+    private boolean hasDeclaredMenuItem() {
+        return hasLoggedIn();
+    }
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -103,6 +121,7 @@ public class ShowVideoDetailActivity extends AppCompatActivity {
             throw new IllegalStateException("Please do intent.putExtra(EXTRA_KEY_VIDEO_ID, vid)");
         }
         mYoukuClientId = getString(R.string.youku_client_id);
+        mVolleySingleton = VolleySingleton.getInstance(this);
 
         //TODO 实现响应式UI
         if (!NetworkChecker.newInstance(this).isConnected()) {
@@ -111,7 +130,7 @@ public class ShowVideoDetailActivity extends AppCompatActivity {
 
         initViews();
 
-        VolleySingleton.getInstance(this).addToRequestQueue(buildVideoDetailRequest(videoIdExtra));
+        mVolleySingleton.addToRequestQueue(buildVideoDetailRequest(videoIdExtra));
     }
 
     private void initViews() {
@@ -131,6 +150,54 @@ public class ShowVideoDetailActivity extends AppCompatActivity {
         super.onDestroy();
     }
 
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        boolean superResult = super.onCreateOptionsMenu(menu);
+        getMenuInflater().inflate(R.menu.menu_activity_show_video_detail, menu);
+        return hasDeclaredMenuItem() || superResult;
+    }
+
+    @Override
+    public boolean onPrepareOptionsMenu(Menu menu) {
+        boolean superResult = super.onPrepareOptionsMenu(menu);
+        MenuItem bookmarkMenuItem = menu.findItem(R.id.action_bookmark);
+        boolean hasLoggedIn = hasLoggedIn();
+        bookmarkMenuItem.setVisible(hasLoggedIn);
+        bookmarkMenuItem.setEnabled(hasLoggedIn);
+        if (mHasBookmarked) {
+            bookmarkMenuItem.setIcon(R.drawable.ic_favorite_white_24dp);
+        } else {
+            bookmarkMenuItem.setIcon(R.drawable.ic_favorite_outline_white_24dp);
+        }
+        return hasDeclaredMenuItem() || superResult;
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        switch (item.getItemId()) {
+            case R.id.action_bookmark:
+                if (!getNetworkChecker().isConnected()) {
+                    NetworkChecker.openNoConnectionDialog(getSupportFragmentManager());
+                    return true;
+                }
+                DefaultSharedPreferencesSingleton defaultSharedPreferences
+                        = getDefaultSharedPreferencesSingleton();
+                if (defaultSharedPreferences.hasLoggedIn()) {
+                    String token = defaultSharedPreferences.retrieveString("access_token", "");
+                    if (mHasBookmarked) {
+                        mVolleySingleton.addToRequestQueue(buildDestroyFavoriteRequest(token));
+                    } else {
+                        mVolleySingleton.addToRequestQueue(buildCreateFavoriteRequest(token));
+                    }
+                } else {
+                    System.err.println("Shouldn't be there.@onOptionsItemSelected.action_bookmark");
+                }
+                return true;
+            default:
+                return super.onOptionsItemSelected(item);
+        }
+    }
+
     private VideoDetailRequest buildVideoDetailRequest(String videoId) {
         final VideoDetailRequestHandler handler =
                 new VideoDetailRequestHandler(new WeakReference<>(this));
@@ -139,6 +206,28 @@ public class ShowVideoDetailActivity extends AppCompatActivity {
                 .setVideo_id(videoId)
                 .setResponseListener(handler)
                 .setErrorListener(handler)
+                .build();
+    }
+
+    /** 创建添加收藏的请求 */
+    private CreateFavoriteRequest buildCreateFavoriteRequest(String token) {
+        return new CreateFavoriteRequest.Builder()
+                .setClient_id(mYoukuClientId)
+                .setVideo_id(mVideo.id)
+                .setAccess_token(token)
+                .setResponseListener(new OnResponseListener(this, RequestType.CREATE_FAVORITE))
+                .setErrorListener(new OnErrorResponseListener(this, RequestType.CREATE_FAVORITE))
+                .build();
+    }
+
+    /** 创建取消收藏的请求 */
+    private DestroyFavoriteRequest buildDestroyFavoriteRequest(String token) {
+        return new DestroyFavoriteRequest.Builder()
+                .setClient_id(mYoukuClientId)
+                .setVideo_id(mVideo.id)
+                .setAccess_token(token)
+                .setResponseListener(new OnResponseListener(this, RequestType.DESTROY_FAVORITE))
+                .setErrorListener(new OnErrorResponseListener(this, RequestType.DESTROY_FAVORITE))
                 .build();
     }
 
@@ -180,6 +269,85 @@ public class ShowVideoDetailActivity extends AppCompatActivity {
             return result;
         }
     }
+
+    private enum RequestType {CREATE_FAVORITE, DESTROY_FAVORITE}
+
+    /**
+     * 此类不会导致内存泄漏
+     */
+    private static class OnResponseListener implements Response.Listener<JSONObject> {
+
+        private final WeakReference<ShowVideoDetailActivity> mActivityWeakReference;
+
+        private final WeakReference<Context> mContextWeakReference;
+
+        private final RequestType mRequestType;
+
+        public OnResponseListener(ShowVideoDetailActivity activity, RequestType requestType) {
+            mActivityWeakReference = new WeakReference<>(activity);
+            mContextWeakReference = new WeakReference<>(activity.getApplicationContext());
+            mRequestType = requestType;
+        }
+
+        @Override
+        public void onResponse(JSONObject jsonObject) {
+            final ShowVideoDetailActivity activity = mActivityWeakReference.get();
+            if (activity != null) {
+                activity.mHasBookmarked = mRequestType == RequestType.CREATE_FAVORITE;
+                activity.invalidateOptionsMenu();
+            }
+            final Context context = mContextWeakReference.get();
+            if (context != null) {
+                Toast.makeText(context, getToastTextResId(), Toast.LENGTH_SHORT).show();
+            }
+        }
+
+        private int getToastTextResId() {
+            switch (mRequestType) {
+                case CREATE_FAVORITE:
+                    return R.string.create_favorite_success;
+                case DESTROY_FAVORITE:
+                    return R.string.destroy_favorite_success;
+                default:
+                    throw new IllegalArgumentException("Unknown RequestType:" + mRequestType);
+            }
+        }
+    }
+
+    /**
+     * 此类不会导致内存泄漏
+     */
+    private static class OnErrorResponseListener implements Response.ErrorListener {
+
+        private final WeakReference<Context> mContextWeakReference;
+
+        private final RequestType mRequestType;
+
+        private OnErrorResponseListener(Context context, RequestType requestType) {
+            mContextWeakReference = new WeakReference<>(context.getApplicationContext());
+            mRequestType = requestType;
+        }
+
+        @Override
+        public void onErrorResponse(VolleyError volleyError) {
+            final Context context = mContextWeakReference.get();
+            if (context != null) {
+                Toast.makeText(context, getToastTextResId(), Toast.LENGTH_SHORT).show();
+            }
+        }
+
+        private int getToastTextResId() {
+            switch (mRequestType) {
+                case CREATE_FAVORITE:
+                    return R.string.create_favorite_failure;
+                case DESTROY_FAVORITE:
+                    return R.string.destroy_favorite_failure;
+                default:
+                    throw new IllegalArgumentException("Unknown RequestType:" + mRequestType);
+            }
+        }
+    }
+
 
     private static class ViewHolder {
