+++ app/src/main/res/menu/menu_activity_show_video_detail.xml
@@ -3,7 +3,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:app="http://schemas.android.com/apk/res-auto"
-        tools:context=".ui.PlayVideoActivity"
+        tools:context=".ui.ShowVideoDetailActivity"
         >
 
     <item
